+++ src/main/java/lotto/model/money/Money.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.money;
 
 public class Money {
