+++ src/main/java/lotto/model/result/LottoRank.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.result;
 
 import java.util.Arrays;
