+++ src/test/java/lotto/model/LottoNumberTest.java
@@ -2,6 +2,7 @@ package lotto.model;
 
 import static org.assertj.core.api.Assertions.*;
 
+import lotto.model.ticket.number.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
