+++ src/test/java/lotto/model/LottoRankTest.java
@@ -2,6 +2,7 @@ package lotto.model;
 
 import static org.assertj.core.api.Assertions.*;
 
+import lotto.model.result.LottoRank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
