+++ src/test/java/lotto/model/LottoStatisticsTest.java
@@ -2,6 +2,9 @@ package lotto.model;
 
 import java.util.List;
 
+import lotto.model.money.Money;
+import lotto.model.result.LottoRank;
+import lotto.model.result.LottoStatistics;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
