+++ src/test/java/lotto/model/WinningTicketTest.java
@@ -3,6 +3,10 @@ package lotto.model;
 import java.util.ArrayList;
 import java.util.List;
 
+import lotto.model.result.LottoRank;
+import lotto.model.ticket.LottoTicket;
+import lotto.model.ticket.WinningTicket;
+import lotto.model.ticket.number.LottoNumber;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
